+++ tests/unit/Support/Entities/Credential.php
@@ -1,0 +1,26 @@
+<?php
+declare(strict_types = 1);
+
+namespace Test\Unit\Support\Entities;
+
+class Credential
+{
+    protected $password;
+    protected $username;
+
+    public function __construct($username, $password)
+    {
+        $this->password = $password;
+        $this->username = $username;
+    }
+
+    public function getPassword()
+    {
+        return $this->password;
+    }
+
+    public function getUserName()
+    {
+        return $this->username;
+    }
+}
