+++ tests/unit/TransformCest.php
@@ -1,12 +1,13 @@
 <?php
 declare(strict_types = 1);
 
-use Diaclone\Resource\ResourceInterface;
+namespace Test\Unit;
+
 use Diaclone\Resource\Collection;
-use Diaclone\Transformer\AbstractTransformer;
-use Diaclone\Transformer\IntegerTransformer;
-use Diaclone\Transformer\StringTransformer;
 use Diaclone\TransformService;
+use Test\Unit\Support\Entities\Person;
+use Test\Unit\Support\Transformers\PersonTransformer;
+use UnitTester;
 
 class TransformCest
 {
@@ -270,113 +271,5 @@ class TransformCest
             ],
         ];
         $I->assertEquals($expected, $output);
-    }
-}
-
-class Person
-{
-    public $my_job;
-
-    protected $friends;
-    protected $name;
-
-    public function __construct($name, $occupation, $friends = [])
-    {
-        $this->friends = $friends;
-        $this->name = $name;
-        $this->my_job = new Occupation($occupation);
-    }
-
-    public function getAge()
-    {
-        return 42;
-    }
-
-    public function getName()
-    {
-        return 'My name is ' . $this->name;
-    }
-
-    public function getNunya()
-    {
-        return 'secret';
-    }
-
-    public function getMyFriends()
-    {
-        return $this->friends;
-    }
-}
-
-class PersonTransformer extends AbstractTransformer
-{
-    protected static $dataTypes = [
-        'my_friends' => Collection::class,
-    ];
-
-    protected static $mappedProperties = [
-        'name'       => 'name',
-        'age'        => 'age',
-        'pigLatin'   => 'pigLatin',
-        'my_job'     => 'occupation',
-        'my_friends' => 'friends',
-    ];
-
-    protected static $transformers = [
-        'age'        => IntegerTransformer::class,
-        'my_job'     => OccupationTransformer::class,
-        'my_friends' => PersonTransformer::class,
-        'pigLatin'   => PigLatinTransformer::class,
-    ];
-}
-
-class Occupation
-{
-    protected $name;
-
-    public function __construct(string $name)
-    {
-        $this->name = $name;
-    }
-
-    public function getName()
-    {
-        return $this->name;
-    }
-}
-
-class OccupationTransformer extends AbstractTransformer
-{
-    protected static $transformers = [
-        'name' => StringTransformer::class,
-    ];
-
-    protected static $mappedProperties = [
-        'name' => 'name',
-    ];
-}
-
-class PigLatinTransformer extends AbstractTransformer
-{
-    public function transform(ResourceInterface $resource)
-    {
-        $value = $this->getPropertyValue($resource->getData(), 'name');
-        $parts = explode(' ', $value);
-        $converted = [];
-        foreach ($parts as $part) {
-            $firstLetter = lcfirst($part[0]);
-            if (in_array($firstLetter, ['a', 'e', 'i', 'o', 'u',])) {
-                $converted[] = $part . 'yay';
-            } else {
-                $pigged = substr($part, 1) . $firstLetter . 'ay';
-                if ($firstLetter !== $part[0]) {
-                    $pigged = ucfirst($pigged);
-                }
-
-                $converted[] = $pigged;
-            }
-        }
-
-        return implode(' ', $converted);
     }
 }
